1.添加文件分享界面及添加邮箱分享功能 2.引入图片库，创建图表界面为写图标做准备

+++ app/src/main/java/com/vitalong/bluetest2/bluepro/SaveDataActivity.java
@@ -340,8 +340,10 @@ public class SaveDataActivity extends AppCompatActivity {
         Utils.createDir(dir);
         File dirFile = new File(dir);
         //删除之前的文件
-        for (String s : dirFile.list()) {
-            new File(dir + s).delete();
+        if (dirFile != null && dirFile.list() != null && dirFile.list().length > 0) {
+            for (String s : dirFile.list()) {
+                new File(dir + s).delete();
+            }
         }
         Date date = new Date();
         SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHHmmss");
